All components in a single file
