--- src/PrimitiveObsession/Bar.cs
@@ -1,4 +1,0 @@
-﻿namespace PrimitiveObsession
-{
-    public class Bar { }
-}
