--- src/PrimitiveObsession/Baz.cs
@@ -1,9 +1,0 @@
-﻿namespace PrimitiveObsession
-{
-    public class Baz
-    {
-        public Baz(Bar bar, string connectionString)
-        {
-        }
-    }
-}
