+++ src/PrimitiveObsession/PrimitiveObsession.csproj
@@ -5,7 +5,7 @@
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{E5C34572-100D-4CAC-9190-73DC5F51688E}</ProjectGuid>
-    <OutputType>Exe</OutputType>
+    <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>PrimitiveObsession</RootNamespace>
     <AssemblyName>PrimitiveObsession</AssemblyName>
@@ -32,6 +32,9 @@
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
+  <PropertyGroup>
+    <StartupObject />
+  </PropertyGroup>
   <ItemGroup>
     <Reference Include="Autofac, Version=4.1.1.0, Culture=neutral, PublicKeyToken=17863af14b0044da, processorArchitecture=MSIL">
       <HintPath>..\packages\Autofac.4.1.1\lib\net45\Autofac.dll</HintPath>
@@ -52,10 +55,7 @@
     <Compile Include="Config\ConnectionString.cs" />
     <Compile Include="IoC\PrimitiveObsessionModule.cs" />
     <Compile Include="Foo.cs" />
-    <Compile Include="Bar.cs" />
-    <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Qux.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
