--- src/PrimitiveObsession/Qux.cs
@@ -1,9 +1,0 @@
-﻿namespace PrimitiveObsession
-{
-    public class Qux
-    {
-        public Qux(Bar bar, string url)
-        {
-        }
-    }
-}
